chore: move unit tests from useCases to tests/unit directory
