--- .github/workflows/integration-tests.yml
@@ -1,57 +1,0 @@
-name: Run Integration Tests
-
-on:
-  pull_request:
-    branches: [main, develop]
-
-jobs:
-  integration-test:
-    runs-on: ubuntu-latest
-
-    services:
-      postgres:
-        image: postgres:15
-        env:
-          POSTGRES_USER: postgres
-          POSTGRES_PASSWORD: postgres
-          POSTGRES_DB: test_db
-        ports:
-          - 5432:5432
-        options: >-
-          --health-cmd pg_isready
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-
-    env:
-      DB_HOST: localhost
-      DB_PORT: 5432
-      DB_USERNAME: postgres
-      DB_PASSWORD: postgres
-      DB_NAME: test_db
-      NODE_ENV: test
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v3
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 22
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Wait for PostgreSQL
-        run: |
-          until pg_isready -h $DB_HOST -p $DB_PORT; do
-            echo "Waiting for PostgreSQL..."
-            sleep 1
-          done
-
-      - name: Run migrations
-        run: npm run typeorm migration:run
-
-      - name: Run integration tests
-        run: npm run test:integration
