+++ package.json
@@ -8,9 +8,10 @@
     "dev": "ts-node-dev --respawn --transpile-only src/server.ts",
     "build": "tsc",
     "start": "node dist/server.js",
-    "test": "cross-env NODE_ENV=test jest",
-    "test:watch": "cross-env NODE_ENV=test jest --detectOpenHandles --watch",
-    "test:cov": "cross-env NODE_ENV=test jest --detectOpenHandles --coverage"
+    "test": "cross-env NODE_ENV=test jest src/tests/unit",
+    "test:watch": "cross-env NODE_ENV=test jest --watch",
+    "test:cov": "cross-env NODE_ENV=test jest --coverage",
+    "test:all": "cross-env NODE_ENV=test jest"
   },
   "repository": {
     "type": "git",
